pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dbslice"
version = "0.0.1"
description = "Minimal DB subsetting CLI bootstrap"
requires-python = ">=3.9"
dependencies = [
    "psycopg[binary]>=3.1,<3.4",
    "PyYAML>=6.0,<7",
]
# psycopg2 is optional at install time for bootstrap; warn at runtime if missing
# Users can `pip install dbslice[psycopg2]` if we later add extras

[project.scripts]
dbslice = "dbslice.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
